fixed favicon toggle bug
